toolkit/profile/test/test_ProfileIdentityUtils.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=543854
-->
<window title="Mozilla Bug 543854"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
   src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
   
    <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=543854"
     target="_blank">Mozilla Test</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[

SimpleTest.waitForExplicitFinish();

const Cc = Components.classes;
const Ci = Components.interfaces;
const Cu = Components.utils;
const Cr = Components.results;

Cu.import("resource:///modules/ProfileIdentityUtils.jsm");

ok(ProfileIdentityUtils.localIdentities[0] == "default"/*"test@example.com"*/,
   "Not the right identity: " + ProfileIdentityUtils.localIdentities[0]);

SimpleTest.finish();

  ]]>
  </script>
</window>
